fix: eliminado el componente RequestCardsContainer para mejorar el mantenimiento a futuro, aunque hay redundancia de código
